Добавляет рефакторинг и исправление ошибок
