+++ js/main.js
@@ -1,1 +1,3 @@
-import './gallery/main';
+import {createImagesData} from './data.js';
+import {setupImageGallery} from './gallery/main.js';
+setupImageGallery(createImagesData());
